Use URL.createObjectURL for reading local files
Somehow I overlooked this method of extracting data from a local file. Now that
we have a real URL for a local file that does not break Cross Origin Request
rules, we can go back to using an `<audio>` tag for playing our media which
greatly simplifies much of the media functionality.