Fix some subtle bugs with the easter egg
Because we were trimming the log to be too long (it was trying to account for
esc), it would only work if the easter egg was the very first string typed.

Additionally, we filter out escape keys explicitly. This accounts for things
like Firefox when it blocks the file input popup, and therefore does not block
swallow the esc.

Also, clean up the array matching logic.

Fixes #395.