Use position fixed to position windows
This ensures that their position does not influence the size of the
window.

Before the scroll size of the browser window was expanding to contain
our webamp windows, which meant we could never measure the natural size
of the window to position ourselves within it.

Thanks to @1j01 for the idea to try `position: fixed;`