Format css according to prettier default config
Looks like I had a .editorconfig in the project root (not sure how it
got there) which was git ignored by my ~/.gitignore. It also turns out
that prettier will respect the indent settings in .editorconfig.

This sets my `.editorconfig` to use the Prettier default, checks it in,
and also reformats all the CSS.

Fixes #684