Implement VerticalScrollbar
This gets rid of rc-slider which was way too heavy, was out of date and I never quiet trusted.

We only ever used a portion of it any way.

Also, this sets us up to have more direct control over the sliders which we will need if we want to support the behavior where you can click once and drag across the EQ sliders to draw a line.