Newly revamped method for basic visualization
All good things are 3...
... Except for Winamp3.

This entirely revamps how Webamp performs its basic visualization (because that is apparently *very* difficult...).

All of this was based on research and decompilation from Winamp 2.63, the way the oscilloscope and spectrum analyzer are drawn, including their different coloring modes (analyzer's Normal, Fire and Line modes, oscilloscope's Dots, Lines and Solid modes) are 100% matched.

The intent behind this new commit and PR is to remove the old way of doing things (VisPainter), while it served us well, maintainability became difficult as time went on (partly because of the way it was written and how additions were made), as things were more based on observation rather than truly fact checking how it actually works.

Fixes the following:
- Polarity of the Oscilloscope is no longer inverted.